Duplicate functions were removed from the backend
